chore: rename pkg wireset -> injection
