+++ internal/app/wire.go
@@ -9,18 +9,18 @@ import (
 	"github.com/google/wire"
 
 	"github.com/bnb-chain/airdrop-service/internal/config"
+	"github.com/bnb-chain/airdrop-service/internal/injection"
 	"github.com/bnb-chain/airdrop-service/internal/module/approval"
 	"github.com/bnb-chain/airdrop-service/internal/module/http"
-	"github.com/bnb-chain/airdrop-service/internal/wireset"
 )
 
 func Initialize(configPath string) (Application, error) {
 	wire.Build(
 		newApplication,
 		config.NewConfig,
-		wireset.InitLogger,
-		wireset.InitKeyManager,
-		wireset.InitStore,
+		injection.InitLogger,
+		injection.InitKeyManager,
+		injection.InitStore,
 		approval.NewApprovalService,
 		http.NewHttpServer,
 	)
