+++ internal/injection/inject_keymanager.go
@@ -1,4 +1,4 @@
-package wireset
+package injection
 
 import (
 	"errors"
